refactor: create a custom hooks to extrat filled logic
