+++ packages/ocean-react/src/TextArea/TextArea.tsx
@@ -1,6 +1,7 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 import classNames from 'classnames';
 
+import useInputFilled from '../_util/useInputFilled';
 import FormControl, { FormControlProps } from '../FormControl';
 
 export type TextAreaProps = Omit<FormControlProps, 'children'> &
@@ -23,25 +24,11 @@ const TextArea = React.forwardRef<HTMLTextAreaElement, TextAreaProps>(
     },
     ref
   ) {
-    const [filled, setFilled] = useState(Boolean(value || defaultValue));
-
-    const handleChange = useCallback(
-      (event) => {
-        if (onChange) return onChange(event);
-
-        // uncontrolled version
-        setFilled(Boolean(event.target.value));
-      },
-      [onChange]
-    );
-
-    useEffect(() => {
-      setFilled(Boolean(value));
-    }, [value]);
-
-    useEffect(() => {
-      setFilled(Boolean(defaultValue));
-    }, [defaultValue]);
+    const { filled, handleChange } = useInputFilled({
+      defaultValue,
+      value,
+      onChange,
+    });
 
     return (
       <FormControl
